ocr-service/requirements.txt
# ocr-service/requirements.txt
fastapi>=0.95.0
uvicorn>=0.17.6
python-multipart>=0.0.5
pdf2image>=1.16.0
Pillow>=9.0.0
# – client OCR, LLM, v.v. nếu có
mistralai>=1.7.0
requests>=2.28.0
python-dotenv>=0.19.0
httpx>=0.22.0